Double meshviewer sampling frequency
There are a lot of missing datapoints visibile in grafana. This could be
related with the expected data point distance for unaggregated data and the
sampling frequency. Both were set to 1min and this could lead to missing
samples. Either because the timers are not perfectly in sync, the sampling
happens slightly after 1 minute or some variable latency in the data
propagation path.