Install workaround for large grafana databases
The grafana 5 database stores a history for all the dashboards. This becomes
large quite fast and will take up all space on the hdd. This for example
caused a 6GB large sqlite3 database which only contained 750KB of useful
information and the rest were newly (1 month?) created backups of the
dashboards.