Create own fastd interface for each domain
Each fastd interface will later be added to a special batX interface. This
is necessary to keep the layer 2 broadcast domains small. The interfaces
are already created (and will be selectable in the firmware domain
selector) to make the transition period slightly less disruptive.

This is done by adding all the new interfaces to the same bat0 interface.
This will be the done this way  until enough nodes transitioned to split
the layer two broadcast domains directly on the servers.