Add fastd interfaces to common bridge
batman-adv has a limited broadcast queue for rebroadcasted packets. This
queue gets full quite fast when a lot of interfaces are added to bat0. This
can be avoided by adding it to a special bridge which doesn't forward
packets between the user ports.

The bridge vpn-ffv will now take all the interfaces from tunneldigger and
from fastd.