Use extra vxlan for BGP and inter-vpn communication
There will be no longer a single batman-adv mesh interface when multidomain
support is implemented on the server. But the BGP communication must still
be exchanged between the servers - a separate/isolated vxlan is the choice
here because the Freifunk clients and APs don't need to be part of this
communication channel.