Switch backup server to external one
The mapdata/statistics server was switched from VPN01 to VPN04. But the
backup server didn't change from VPN04 to something else.

It doesn't make sense to backup the data from VPN04 to VPN04. And there is
no other official freifunk-vogtland.net server available with enough
storage capacity to save the backups. An external server is therefore used
for it.