Use ip-link to create/destroy batadv interfaces
There seems to be an odd problem where the batadv interface is created on
startup but the main bridge slave is not actually added. The reason for
this is unknown and it has to be debugged how this can happen.

Unfortunately, batctl combines all the steps to create and add an interface
in the same command. So it might be easier to use specialized commands and
then debug each command separately.