Delay services with dependencies to networkd configured interfaces
systemd-networkd tends to remove links between interfaces (slave/master,
...) when it starts to configure them. But systemd was already informed
that network.target and sys-subsystem-net-devices-* already finished and
started them accordingly.

For example, the vpn bridge could have been started and added successfully
to the batadv interface as slave. Some milliseconds later, systemd-networkd
continues to configure the vpn bridge and removes it from the batadv
interface again.

Something similar was also observed with tunneldigger interface which were
added to the vpn-bridge but it is unknown whether this was actually caused
by systemd-networkd.