Add scheduler do balance fastd connections between servers
The servers have most of the time a not balanced number of connections.
Limiting the fastd connections more using static limits causes problems
when the number of nodes with internet connections increases over this
limit.

The scheduler tries to reduce this kind of problem by automatically
increasing the limit based on the number of fastd connections on all
servers. The number of connections on a server can also be (relatively)
adjusted by increasing its specific cost for a single fastd connection to
this server. The cost sum (clients * cost) is then tried to be kept
balanced between all servers.

It is also possible to limit some servers to an hard fastd connection limit
(per domain). If this limit is reached, this server is not used anymore in
the balancing calculation.