Use haveged to work around missing hwrng for Linux 4.19+
Linux 4.19 doesn't accept the systemd-random-seed.service as input for
randomness. So it easily happens that a server takes half an hour to gather
enough randomness to start up SSH or any other server which uses randomness
to start its encryption modules.

The best solution would be to get a hwrng generator or use virtio-rng*
exposed to the server. Unfortunately, this doesn't seem to be the case at
the moment for most of the existing servers.