Move IPv4 and IPv6 Traffic over VPN03 instead of VPN05
Larger Packets (ping 1229 -> 1271 Ethernet packets) are currently dropped
for unknown reasons over the GRE tunnels when the traffic goes over the
linevast infrastructure.

So it is better for us to use VPN03 for the GRE tunnels than VPN05. This is
explicitely done for IPv4 but the IPv6 traffic will be done imlicitely at
the moment because we don't announce a route over RADV.

Signed-off-by: Sven Eckelmann <sven@narfation.org>