Move firmware.freifunk-vogtland.net to vpn01
The build01 server was not used for actual build jobs since a while. Only
the firmware.freifunk-vogtland.net HTTP server was realized over it. But it
never received a public IPv4 address and thus the HTTP(S) conneciton was
terminated vpn01 and a new proxy connection was established by the vpn01
webserver.

This hack can be dropped when vpn01 is directly providing the files. And
the HTTP(S) port(s) for the old build01 IPv6 address are now automatically
redirected to the vpn01 server to allow older firmware versions to still
upgrade without a reachable DNS server.

Signed-off-by: Sven Eckelmann <sven@narfation.org>