radvd: Reduce lifetime of lease
APs are used to forward DNS requests to servers. And they are using IPv6
for talk to the actual VPN servers. If such an VPN server just disappears,
a couple of the nodes will still have the disappeared VPN server as their
preferred route (+their prefix as their preferred address). In this case,
the DNS requests for this server will disappear into nowhere and clients
will not be able to resolve domain names anymore.

By default, the preferred_lft is a whole day. Reduce this to just 5 minutes
to don't break the network for a day when a server stops working.

Signed-off-by: Sven Eckelmann <sven@narfation.org>