firewall: Ensure ICMPv6 replies are returned correctly via FFRL
When some requests come from Freifunk Rheinland (Backbone) to our servers
and an error has to be returned then this error was always returned via the
default route. And the default route was not responsible for the FFRL
specific IPv6 address.

So addresses which have as source the FFRL bb-* specific IPv6 address must
always be send via the freifunk table (which contains the FFRL specific
routes).

This was noticed when Linus ran a traceroute6 (with mtu) from an external
IP range to a FFV specific IP address:

  $ traceroute6 --mtu 2a03:2260:200f:101:ae86:74ff:fec7:d100
  traceroute to 2a03:2260:200f:101:ae86:74ff:fec7:d100 (2a03:2260:200f:101:ae86:74ff:fec7:d100), 30 hops max, 65000 byte packets
   1  (anonymized for this commit message)
   2  2a02:3001::123 (2a02:3001::123)  60.127 ms F=1492  13.594 ms  12.937 ms
   3  2a02:3001::d1 (2a02:3001::d1)  12.642 ms  11.921 ms  11.859 ms
   4  2a02:3001::216 (2a02:3001::216)  12.070 ms  12.374 ms  12.582 ms
   5  2a03:2260::4 (2a03:2260::4)  11.744 ms  12.425 ms  11.853 ms
   6  * F=1400 * *
   7  * * *
   8  * * *
   9  * * *
  10  * * *
  11  * * *
  12  *^C

On the other end, it was visible that the too large packet came in as
expected but the ICMPv6 reply was transmitted via the default route:

  $ sudo tcpdump -ni any 'host 2a00::abcd'
  19:26:19.365741 bb-b-ak-ber In  IP6 2a00::abcd.58155 > 2a03:2260:200f:101:ae86:74ff:fec7:d100.33507: UDP, length 1352
  19:26:19.365911 eth1  Out IP6 2a03:2260:0:417::2 > 2a00::abcd: ICMP6, packet too big, mtu 1280, length 1240

So just hack in the "source specific" route selection via a mark that then
forces the routing rules to select the freifunk specific table.

We might want to move this later to ip route.

Reported-by: Linus Lüssing <linus.luessing@c0d3.blue>
Signed-off-by: Sven Eckelmann <sven@narfation.org>