fastd: Disable client connections on vpn02
The VPN02 has problems with its router which seems to increase the ping to
any device in the same network (as long as its on the routers internal
switch) extremely. This only happens sporadically and can currently be
explained by some (yet unidentified) bug in the router.

This problem seems to happen especially when the VPN02 transfers data over
vxlan. And since the router is also used for private connectivity, the
vxlan was deactivated all the time. Which also means that the connectivity
to all other FFV VPN servers was dropped - and since VPN02 doesn't have an
own link to the FFRL backbone, this broke the network connectivity for each
AP (and all its clients) which were connected to this server.

So just disable fastd on this server for now until another solution was
found.

Signed-off-by: Sven Eckelmann <sven@narfation.org>