vxlan: Switch to IPv6 based vxlan destinations
The router problems on vpn02 might be related to some conntracking
timeout/overflow problem in the FritzBox. So just try IPv6 for VXLAN
communication to see if it works better. Unfortunately, this also reduces
the MTU from 1450 (for vxlan0) to 1430.

The port had to changed to 4790 because the UDP port 4789 (IPv6) is alrady
used for mow-ffv_pl and unfortunately not completely bound to its device
enp2s0.

Signed-off-by: Sven Eckelmann <sven@narfation.org>